App-FaceRecognition-main/requirements.txt
asgiref==3.7.2
boto3==1.34.79
botocore==1.34.79
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
Django==5.0.1
django-crispy-forms==2.1
django-storages==1.14.2
djangorestframework==3.15.1
jmespath==1.0.1
pillow==10.2.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
